Refactor robot to load the texture explicit
Refactor robot to load the texture explicit when the gamescreen is loaded and not when the robot is initialized. This makes it so the test are green (had nullpointer exception before, because it expected a full game initialized).
And used Intellj to reformat code (ctrl+alt+L). This added some whitespaces.