Refactor of the code from assignment 1 (#24)
* Updated the player texture to be more expressive

Updated the player texture to be more expressive
and removed the outlier lines

* Refactor loading of texture regions into own function

* reformat and removed test specific constructor

* Extrated large code blocks to methods

* Refactor robot to load the texture explicit

Refactor robot to load the texture explicit when the gamescreen is loaded and not when the robot is initialized. This makes it so the test are green (had nullpointer exception before, because it expected a full game initialized).
And used Intellj to reformat code (ctrl+alt+L). This added some whitespaces.